spritualroot/pyproject.toml
[tool.poetry]
name = "spritualroot"
version = "0.1.0"
description = "Impact NGO Application"
authors = ["Shivam Chaurasia <SHIVAM9172@GMAIL.COM>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
django-thumbnails = "^0.7.0"
django-ckeditor = "^6.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
